Create a file for each cli command.
