+++ cmd/storage.go
@@ -1,0 +1,9 @@
+package cmd
+
+import "github.com/spf13/cobra"
+
+func NewStorage() *cobra.Command {
+	return &cobra.Command{
+		Use: "storage",
+	}
+}
